+++ src/styles/global.ts
@@ -12,7 +12,7 @@ export default createGlobalStyle`
   #root {
     max-width: 700px;
     margin: 0 auto;
-    padding: 40px 20px;
+    padding: 20px;
   }
 
   html {