+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2022-05-04
+### Added
+- Home Page: Added pagination for the repository list.
+- Repository issues component: Show pagination only if have more issues than items per page (10).
+
+### Fixed
+- Pagination component: Changed pagination info to start by register 1 instead of 0.
+
 ## [0.3.0] - 2022-05-04
 ### Added
 - Pagination component.